adding to api and creatign public page
