+++ packages/geoview-core/public/index.html
@@ -31,6 +31,7 @@
     <h4>Basic Raw Components</h4>
     <a class="page-link" href="./raw-feature-info.html">Basic Feature Info</a>
     <a class="page-link" href="./raw-data-table.html">Basic Data Table</a>
+    <a class="page-link" href="./layer-select.html">Layer Select</a>
     <h4>Basic Loading</h4>
     <a class="page-link" href="./default-config.html">Default Configuration</a>
     <a class="page-link"
