+++ packages/geoview-core/public/templates/layer-select.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>
+    <%= htmlWebpackPlugin.options.title %>
+  </title>
+  <link rel="shortcut icon" href="./favicon.ico" />
+  <meta name="msapplication-TileColor" content="#da532c" />
+  <meta name="msapplication-config" content="./img/browserconfig.xml" />
+  <meta name="theme-color" content="#ffffff" />
+  <meta name="msapplication-TileColor" content="#da532c" />
+  <meta name="theme-color" content="#ffffff" />
+  <link href="https://fonts.googleapis.com/css?family=Roboto|Montserrat:200,300,400,900|Merriweather" rel="stylesheet"
+    type="text/css" />
+  <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
+  <link rel="stylesheet" href="css/style.css" />
+</head>
+
+<body>
+  <div class="header-table">
+    <table>
+      <tbody>
+        <tr>
+          <td><img class="header-logo" alt="logo" src="./img/Logo.png" /></td>
+          <td class="header-title">
+            <h1><strong>Package - Raw Data Table</strong></h1>
+          </td>
+        </tr>
+        <tr>
+          <td>
+            <a href="./index.html">Main</a><br />
+          </td>
+        </tr>
+      </tbody>
+    </table>
+  </div>
+
+  <div class="map-title-holder">
+    <h4 id="HUC1">1. Default Configuration</h4>
+    <a class="ref-link" href="#top">Top</a>
+  </div>
+  <div id="mapWM1" class="llwp-map" data-lang="en" data-config-url="./configs/raw-data-table.json"></div>
+
+  <div>
+    <button onClick="renderLayerSelectComponent()" id="createDataTableFromMapBtn">Draw Layer Select Component
+    </button>
+
+    <div id="dataTable" style="margin: 1rem 0"></div>
+  </div>
+
+
+  <script src="codedoc.js"></script>
+  <script>
+    // initialize cgpv and api events, a callback is optional, used if calling api's after the rendering is ready
+    cgpv.init(function () {
+      console.log('api is ready');
+      const { react, createRoot, api, ui } = cgpv;
+
+      //create snippets
+      createConfigSnippet();
+      createCodeSnippet();
+
+      renderLayerSelectComponent = async () => {
+        console.time("MapTableCreation")
+        const layers = Object.keys(api.maps['mapWM1'].layer.geoviewLayers);
+        const itemAcc = {
+          title: "Layers Select",
+          content: await cgpv.api.maps['mapWM1'].layersSelect.createLayersSelect({layerIds: layers, mapId: "mapWM1", isRemovable: false })
+        }
+
+        cgpv.api.addUiComponent(
+          "dataTable",
+          cgpv.react.createElement(cgpv.ui.elements.Accordion, { items: [itemAcc] })
+        );
+        console.timeEnd("MapTableCreation")
+      }
+    });
+  </script>
+</body>
+
+</html>
