+++ packages/geoview-core/src/core/components/layers-select/types.ts
@@ -2,6 +2,18 @@ import { TypeLayerEntryConfig } from '@/geo/map/map-schema-types';
 
 import { AbstractGeoViewLayer } from '@/geo/layer/geoview-layers/abstract-geoview-layers';
 
+export interface LayersSelectProps {
+  mapId: string;
+  layerIds: string[];
+  isRemoveable: false;
+  canSetOpacity?: boolean;
+  expandAll?: boolean;
+  hideAll?: boolean;
+  canZoomTo?: boolean;
+  canSort?: boolean;
+  onOpenDetails?: (layerId: string, layerConfigEntry: TypeLayerEntryConfig | undefined) => void;
+}
+
 export interface LayerSelectItemProps {
   layerId: string;
   geoviewLayerInstance: AbstractGeoViewLayer;
